+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- `Zoi.non_negative/1` refinement for numbers to accept values from 0 and above
+- `Zoi.non_negative/2` refinement for numbers to accept values from 0 and above
 
 ## 0.8.0 - 2025-10-26